# src/main/resources/META-INF/resources/schema/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "assoparenteleve API",
    "version" : "1.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ],
...
  "paths" : {
...
    "/commande/evenement/{id}" : {
      "get" : {
        "tags" : [ "Commande" ],
        "summary" : "Get an commande by ID",
        "description" : "Returns a single commande",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Commande"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Commande not found"
          }
        }
      }
    },
...
    "/evenement" : {
      "get" : {
        "tags" : [ "Evenement" ],
        "summary" : "List all evenements",
        "description" : "Returns a list of all evenements",
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Evenement"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "No evenements found"
          }
        }
      },
      "put" : {
        "tags" : [ "Evenement" ],
        "summary" : "Update an evenement",
        "description" : "Returns the updated evenement",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Evenement"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Evenement updated successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Evenement"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      },
      "post" : {
        "tags" : [ "Evenement" ],
        "summary" : "Create a new evenement",
        "description" : "Returns the created evenement",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Evenement"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Evenement created successfully"
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    },
    "/evenement/{id}" : {
      "get" : {
        "tags" : [ "Evenement" ],
        "summary" : "Get an evenement by ID",
        "description" : "Returns a single evenement",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Evenement"
                }
              }
            }
          },
          "404" : {
            "description" : "Evenement not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "Evenement" ],
        "summary" : "Delete an evenement",
        "description" : "Deletes an evenement by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Evenement deleted successfully"
          },
          "404" : {
            "description" : "Evenement not found"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Adulte" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "nom" : {
            "type" : "string"
          },
          "prenom" : {
            "type" : "string"
          },
          "mail" : {
            "type" : "string"
          },
          "role" : {
            "$ref" : "#/components/schemas/Role"
          },
          "motDePasse" : {
            "type" : "string"
          },
          "telephone" : {
            "type" : "string"
          },
          "eleves" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Eleve"
            }
          },
          "active" : {
            "type" : "boolean"
          }
        }
      },
      "Commande" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "payer" : {
            "type" : "boolean"
          },
          "distribuer" : {
            "type" : "boolean"
          },
          "evenement" : {
            "format" : "int32",
            "type" : "integer"
          },
          "adulte" : {
            "$ref" : "#/components/schemas/Adulte"
          },
          "eleve" : {
            "$ref" : "#/components/schemas/Eleve"
          },
          "commandeProduits" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CommandeProduit"
            }
          }
        }
      },
      "CommandeProduit" : {
        "type" : "object",
        "properties" : {
          "produit" : {
            "$ref" : "#/components/schemas/Produit"
          },
          "quantite" : {
            "format" : "int32",
            "type" : "integer"
          }
        }
      },
      "Distribution" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "distributeur" : {
            "type" : "string"
          },
          "lieu" : {
            "type" : "string"
          },
          "dateDistribution" : {
            "type" : "string"
          }
        }
      },
      "Eleve" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "nom" : {
            "type" : "string"
          },
          "prenom" : {
            "type" : "string"
          },
          "dateNaissance" : {
            "type" : "string"
          },
          "niveau" : {
            "type" : "string"
          }
        }
      },
      "Evenement" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "nom" : {
            "type" : "string"
          },
          "dateDebut" : {
            "type" : "string"
          },
          "dateFin" : {
            "type" : "string"
          },
          "datePaiement" : {
            "type" : "string"
          },
          "distributionList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Distribution"
            }
          },
          "produitList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Produit"
            }
          },
          "valider" : {
            "type" : "boolean"
          },
          "avecEleve" : {
            "type" : "boolean"
          }
        }
      },
      "Produit" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "nom" : {
            "type" : "string"
          },
          "fournisseur" : {
            "type" : "string"
          },
          "prix" : {
            "format" : "double",
            "type" : "number"
          },
          "option" : {
            "type" : "boolean"
          }
        }
      },
      "Role" : {
        "enum" : [ "SUPER_ADMIN", "ADMIN", "PARENT", "PROFESSOR" ],
        "type" : "string"
      }
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "http",
        "description" : "Authentication",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}